[IMP] mail: unskip "composer state conserved when clicking on another topbar button"
