+++ addons/mail/static/tests/qunit_suite_tests/components/chatter_topbar_tests.js
@@ -1,11 +1,6 @@
 /** @odoo-module **/
 
-import {
-    afterNextRender,
-    nextAnimationFrame,
-    start,
-    startServer,
-} from "@mail/../tests/helpers/test_utils";
+import { afterNextRender, start, startServer } from "@mail/../tests/helpers/test_utils";
 
 import { makeTestPromise } from "web.test_utils";
 
@@ -56,67 +51,5 @@ QUnit.module("mail", {}, function () {
                 "attachments button should now have a loader"
             );
         });
-
-        QUnit.skipRefactoring(
-            "composer state conserved when clicking on another topbar button",
-            async function (assert) {
-                assert.expect(8);
-
-                const pyEnv = await startServer();
-                const resPartnerId1 = pyEnv["res.partner"].create({});
-                const { click, openView } = await start();
-                await openView({
-                    res_id: resPartnerId1,
-                    res_model: "res.partner",
-                    views: [[false, "form"]],
-                });
-
-                assert.containsOnce(
-                    document.body,
-                    `.o-mail-chatter-topbar`,
-                    "should have a chatter topbar"
-                );
-                assert.containsOnce(
-                    document.body,
-                    `button:contains(Send message)`,
-                    "should have a send message button in chatter menu"
-                );
-                assert.containsOnce(
-                    document.body,
-                    `button:contains(Log note)`,
-                    "should have a log note button in chatter menu"
-                );
-                assert.containsOnce(
-                    document.body,
-                    `button[aria-label='Attach files']`,
-                    "should have an attachments button in chatter menu"
-                );
-
-                await click(`button:contains(Log note)`);
-                assert.containsOnce(
-                    document.body,
-                    `button:contains(Log note).o-active`,
-                    "log button should now be active"
-                );
-                assert.containsNone(
-                    document.body,
-                    `button:contains(Send message)`,
-                    "send message button should not be active"
-                );
-
-                $(`button[aria-label='Attach files']`)[0].click();
-                await nextAnimationFrame();
-                assert.containsOnce(
-                    document.body,
-                    `button:contains(Log note).o-active`,
-                    "log button should still be active"
-                );
-                assert.containsNone(
-                    document.body,
-                    `button:contains(Send message).o-active`,
-                    "send message button should still be not active"
-                );
-            }
-        );
     });
 });
